### 2019-09-06 #1713 @arfon
:wave: @mdc007, many thanks for your submission to JOSS. Firstly, this pull request should fix your paper compilation: https://github.com/matt002/GaitPy/pull/1

Second, it's not immediately clear to me that your submission is currently up to standard for JOSS yet I'm afraid. In particular the fact that this is a written for Python 2.7 suggests to me that this package is close to deprecated and not ['designed for maintainable extension'](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements).

Also, there don't currently appear to be any automated/unit tests which will be a problem during review.
