+++ gaitpy/model/feature_order.txt
@@ -1,7 +1,5 @@
-lumbar_y_bp_filt_[0.5, 3.0]_dom_freq_magnitude
 lumbar_y_bp_filt_[0.5, 3.0]_dom_freq_ratio
 lumbar_y_bp_filt_[0.5, 3.0]_dom_freq_value
-lumbar_y_bp_filt_[0.5, 3.0]_iqr_of_autocovariance
 lumbar_y_bp_filt_[0.5, 3.0]_mean_cross_rate
 lumbar_y_bp_filt_[0.5, 3.0]_range
 lumbar_y_bp_filt_[0.5, 3.0]_rms
