### 2019-09-17 1df8344
minor updates to paper
