### 2019-09-18 f02674e
update MANIFEST.in
