### 2019-10-02 #1713 @arfon
@finsberg @abhishektha - would you both be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
