### 2019-10-02 #1778 @whedon
[REVIEW]: GaitPy: An Open-Source Python Package for Gait Analysis Using an Accelerometer on the Lower Back
**Submitting author:** @matt002 (<a href="http://orcid.org/0000-0001-9954-7003">Matthew Czech</a>)
**Repository:** <a href="https://github.com/matt002/GaitPy" target ="_blank">https://github.com/matt002/GaitPy</a>
**Version:** v1.0
