### 2019-10-04 98f917d
resolved issues
- Removed .pyc files
- updated setup.py
- fixed .gitignore
