### 2019-10-04 83e0fc9
Delete .gitignore.txt
