+++ .gitignore.txt
@@ -1,9 +1,0 @@
-*.xml
-*.iml
-*.pyc
-.DS_Store
-/dist/
-/*.egg-info
-.idea/
-.eggs/
-build/
