### 2019-10-07 9329166
Update MANIFEST.in - add missing file gaitpy/demo/demo_gait_features.csv
