+++ docs/source/index.rst
@@ -1,0 +1,13 @@
+.. GaitPy documentation master file, created by
+   sphinx-quickstart on Mon Oct 14 13:09:53 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to GaitPy's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 4
+   :caption: Contents:
+
+   modules
