+++ docs/source/gaitpy.rst
@@ -1,0 +1,12 @@
+GaitPy
+-----------------------
+
+.. automodule:: gaitpy.gait.Gaitpy
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. toctree::
+   :maxdepth: 4
+
+   gaitpy_functions
