+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+Modules
+---------
+
+.. toctree::
+   :maxdepth: 4
+
+   gaitpy
