+++ docs/html/_sources/gaitpy_functions.rst.txt
@@ -1,0 +1,23 @@
+classify_bouts
+-----------------------
+
+.. automodule:: gaitpy.gait.Gaitpy.classify_bouts
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+extract_features
+-----------------------
+
+.. automodule:: gaitpy.gait.Gaitpy.extract_features
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+plot_contacts
+-----------------------
+
+.. automodule:: gaitpy.gait.Gaitpy.plot_contacts
+    :members:
+    :undoc-members:
+    :show-inheritance:
