### 2019-10-15 4aed511
documentation updates
