+++ .travis.yml
@@ -1,0 +1,3 @@
+language: python
+python:
+ - "3.6"
