### 2019-10-18 matt002/GaitPy#9 @finsberg
[JOSS REVIEW] Troubles with installation
I am having a hard time installing this on Mac OSX.

1. Do you need to pin your dependencies? If not I would recommend not to pin them.

2. I was not able to install this on Mac using python setup.py install. However, I was able to do it using pip (pip install .), and only using non-pinned dependencies.

3. The package on pypi does not include the latest version (with `setup_requires`)
