+++ setup.py
@@ -12,12 +12,12 @@ setuptools.setup(name='gaitpy',
                  description='Read and process raw vertical accelerometry data from a lumbar sensor during gait; calculate clinical gait characteristics.',
                  long_description=long_description,
                  long_description_content_type="text/markdown",
-                 url='http://github.com/PfizerRD/gaitpy',
+                 url='http://github.com/matt002/gaitpy',
                  packages=setuptools.find_packages(),
                  author='Matthew Czech',
                  author_email='czech1117@gmail.com',
                  keywords=['gait', 'lumbar', 'sensor', 'digital', 'wearable', 'python', 'inverted pendulum', 'czech'],
-                 classifiers=["Programming Language :: Python :: 2.7",
+                 classifiers=["Programming Language :: Python :: 3.6",
                               "License :: OSI Approved :: MIT License"],
                  license='MIT',
                  zip_safe=False,
