### 2019-10-24 a2d21c1
unit testing updates
