+++ tests/test_util.py
@@ -1,0 +1,166 @@
+import os
+import pandas as pd
+from pandas.util.testing import assert_frame_equal
+import warnings
+warnings.simplefilter(action='ignore', category=FutureWarning)
+import numpy as np
+import gaitpy.util as util
+from gaitpy.gait import *
+
+def test_load_data():
+    # Load/format data
+    src = os.path.abspath(__file__ + '/../../')+'/gaitpy/demo/demo_data.csv'
+    raw_data = pd.read_csv(src, skiprows=99, names=['timestamps', 'x', 'y', 'z'], usecols=[0, 1, 2, 3])
+    raw_data['unix_timestamps'] = pd.to_datetime(raw_data.timestamps, format="%Y-%m-%d %H:%M:%S:%f").values.astype(np.int64) // 10**6
+    raw_data = raw_data.iloc[:10,:]
+
+    # Create an instance of GaitPy
+    gaitpy = Gaitpy(raw_data,
+                    50,
+                    v_acc_col_name='y',
+                    ts_col_name='unix_timestamps',
+                    v_acc_units='g',
+                    ts_units='ms',
+                    flip=False)
+
+    # Run function being tested
+    obtained_y_accel, obtained_ts = util._load_data(gaitpy, gaitpy.down_sample)
+
+    # Confirm expected results
+    expected_y_accel = pd.Series(np.array([7.138261,7.177487,7.177487,7.215733,6.177209,7.868856,7.676646,5.792788,4.831736,10.713765]), name='y')
+    pd.testing.assert_series_equal(obtained_y_accel, expected_y_accel)
+
+    expected_ts = pd.Series(np.array([1565087150000,1565087150020,1565087150040,1565087150060,1565087150080,
+                                      1565087150100,1565087150120,1565087150140,1565087150160,1565087150180]), name='unix_timestamps')
+    pd.testing.assert_series_equal(obtained_ts, expected_ts)
+
+def test_extract_signal_features():
+    # Load/format data
+    src = os.path.abspath(__file__ + '/../../')+'/gaitpy/demo/demo_data.csv'
+    raw_data = pd.read_csv(src, skiprows=99, names=['timestamps', 'x', 'y', 'z'], usecols=[0, 1, 2, 3])
+    raw_data['unix_timestamps'] = pd.to_datetime(raw_data.timestamps, format="%Y-%m-%d %H:%M:%S:%f").values.astype(np.int64) // 10**6
+    data = pd.DataFrame({'y': raw_data.iloc[:150,:].y})
+    timestamps = pd.DatetimeIndex(raw_data.iloc[:150,:].unix_timestamps.astype('datetime64[ms]'))
+
+    # Run function being tested
+    obtained_feature_set, obtained_start_times_list, obtained_end_times_list = util._extract_signal_features(data, timestamps, 50)
+
+    # Confirm expected results
+    expected_feature_set = pd.DataFrame([[2.67261662, 0.04452866, 0.19341188, 0.5859375, 0.28124919, 0.65097933, -12.20492315, 0.59233586, 0.04666667]],
+                                        columns=['lumbar_y_bp_filt_[0.5, 3.0]_signal_entropy','lumbar_y_bp_filt_[0.5, 3.0]_rms',
+                                                     'lumbar_y_bp_filt_[0.5, 3.0]_range','lumbar_y_bp_filt_[0.5, 3.0]_dom_freq_value',
+                                                     'lumbar_y_bp_filt_[0.5, 3.0]_dom_freq_magnitude','lumbar_y_bp_filt_[0.5, 3.0]_dom_freq_ratio',
+                                                     'lumbar_y_bp_filt_[0.5, 3.0]_spectral_flatness','lumbar_y_bp_filt_[0.5, 3.0]_spectral_entropy',
[…]
