### 2019-10-24 dd565b1
unit test updates
