+++ requirements.txt
@@ -1,9 +1,9 @@
-deepdish==0.3.4
+deepdish
+pandas
+scipy
+numpy
+PyWavelets
+scikit-learn
+statsmodels
 bokeh==0.12.10
-pandas==0.20.3
-scipy==1.2.0
-numpy==1.13.3
-PyWavelets==0.5.2
-scikit-learn==0.18.2
 dill==0.2.7.1
-statsmodels==0.8.0
