+++ requirements.txt
@@ -1,9 +1,9 @@
-deepdish
-pandas
-scipy
-numpy
-PyWavelets
-scikit-learn
-statsmodels
+pandas==0.20.3
+scipy==1.2.0
+numpy==1.13.3
+PyWavelets==0.5.2
+scikit-learn==0.18.2
+statsmodels==0.8.0
 bokeh==0.12.10
 dill==0.2.7.1
+deepdish==0.3.4
