### 2019-10-25 matt002/GaitPy#9 @matt002
It is very important in my area of research that the methods/algorithms used are consistent going forward. I have experimented with unpinning dependencies, however I have gotten unit-test errors. Would it be possible to let me know what errors/dependencies are causing the setup.py installation issue? Also, I have uploaded the most recent version to pypi.
