+++ docs/source/index.rst
@@ -36,23 +36,20 @@ Basic usage
 -----------
 Gaitpy consists of the following 3 functions:
 
-1. classify_bouts: If your data consists of gait and non-gait data, run the classify_bouts function to first
-classify bouts of gait. If your data is solely during gait, this function is not necessary to use.
-2. extract_features: Extract initial contact (IC) and final contact (FC) events from your data and estimate
-various temporal and spatial gait features.
-3. plot_contacts: Plot the resulting bout detections and IC/FC events alongside your raw accelerometer signal.
+	1. classify_bouts: If your data consists of gait and non-gait data, run the classify_bouts function to first classify bouts of gait. If your data is solely during gait, this function is not necessary to use.
+	2. extract_features: Extract initial contact (IC) and final contact (FC) events from your data and estimate various temporal and spatial gait features.
+	3. plot_contacts: Plot the resulting bout detections and IC/FC events alongside your raw accelerometer signal.
 
 Gaitpy accepts a csv file or pandas dataframe that includes a column containing unix timestamps and a column containing
 vertical acceleration from a lumbar-mounted sensor. Gaitpy makes three assumptions by default:
 
-1. Timestamps and vertical acceleration columns are labeled 'timestamps' and 'y' respectively, however
-this can be changed using the 'v_acc_col_name' and 'ts_col_name' arguments respectively.
-2. Timestamps are in Unix milliseconds and data is in meters per second squared, however this can be be changed using the 'ts_units' and 'v_acc_units' arguments respectively.
-3. Baseline vertical acceleration data is -9.8m/s^2 or -1g. If your baseline data is +9.8m/s^2 or +1g, set the 'flip' argument to True.
+	1. Timestamps and vertical acceleration columns are labeled 'timestamps' and 'y' respectively, however this can be changed using the 'v_acc_col_name' and 'ts_col_name' arguments respectively.
+	2. Timestamps are in Unix milliseconds and data is in meters per second squared, however this can be be changed using the 'ts_units' and 'v_acc_units' arguments respectively.
+	3. Baseline vertical acceleration data is -9.8m/s^2 or -1g. If your baseline data is +9.8m/s^2 or +1g, set the 'flip' argument to True.
 
 Additionally, the sample rate of your device (at least 50Hz) and height of the subject must be provided.
 
-More details about the inputs and ouputs of each of these functions can be found in Czech et al. 2019.[![status](https://joss.theoj.org/papers/a2233c9e27db0b6625dc56a3f7363875/status.svg)](https://joss.theoj.org/papers/a2233c9e27db0b6625dc56a3f7363875)
+More details about the inputs and ouputs of each of these functions can be found in `Czech et al. 2019 (in preparation) <https://joss.theoj.org/papers/a2233c9e27db0b6625dc56a3f7363875>`_.
 
 .. code::
 
