### 2019-10-25 matt002/GaitPy#9 @finsberg
I see. I think if you managed to get the continuous integration system working, and everything installs correctly there, it is probably OK. One thing that would be very useful would be if you could create a conda package (see https://github.com/matt002/GaitPy/issues/6). Alternatively you could write in the installation instructions that it is only tested on Linux.
