### 2019-10-25 3ed2392
updated docs
