+++ gaitpy/gait.py
@@ -6,34 +6,41 @@ class Gaitpy():
     - Machine learning based method of bout classification.
     - Visualizing results.
 
-    data: str or pandas.core.frame.DataFrame
-        - Option 1: Pandas dataframe containing unix timestamp column and vertical acceleration data during gait, both of type float
-        - Option 2: File path of .csv file containing timestamp column and vertical acceleration data during gait. One column should contain unix timestamps of type float -- by default gaitpy will assume the column title is 'timestamps' with units in milliseconds. A second column should be vertical acceleration of type float -- by default gaitpy will assume the column title is 'y' with units in m/s^2.
+    Parameters:
+        data: str or pandas.core.frame.DataFrame
+            - Option 1: Pandas dataframe containing unix timestamp column and vertical acceleration data during gait, both of type float
 
-    sample_rate: int or float
-        Sampling rate of accelerometer data in Hertz.
+            - Option 2: File path of .csv file containing timestamp column and vertical acceleration data during gait. One column should contain unix timestamps of type float (by default gaitpy will assume the column title is 'timestamps' with units in milliseconds). A second column should be vertical acceleration of type float (by default gaitpy will assume the column title is 'y' with units in m/s^2).
 
-    v_acc_col_name: str
-        Column name of the vertical acceleration data ('y' by default)
+        sample_rate: int or float
+            Sampling rate of accelerometer data in Hertz.
 
-    ts_col_name: str
-        Column name of the timestamps ('timestamps' by default)
+        v_acc_col_name: str
+            Column name of the vertical acceleration data ('y' by default)
 
-    v_acc_units: str
-        Units of vertical acceleration data ('m/s^2' by default). Options:
-        - 'm/s^2' = meters per second squared
-        - 'g' = standard gravity
+        ts_col_name: str
+            Column name of the timestamps ('timestamps' by default)
 
-    ts_units: str
-        Units of timestamps ('ms' by default). Options:
-        - 's' = seconds
-        - 'ms' = milli-seconds
-        - 'us' = microseconds
+        v_acc_units: str
+            Units of vertical acceleration data ('m/s^2' by default). Options:
 
-    flip: bool
-        Boolean specifying whether to flip vertical acceleration data before analysis (False by default). Algorithm
-        assumes that baseline vertical acceleration data is at -9.8 m/s^2 or -1g. (ie. If baseline data in vertical
-        direction is 1g, set 'flip' argument to True)
+            - 'm/s^2' = meters per second squared
+
+            - 'g' = standard gravity
+
+        ts_units: str
+            Units of timestamps ('ms' by default). Options:
+
+            - 's' = seconds
+
+            - 'ms' = milli-seconds
+
+            - 'us' = microseconds
+
+        flip: bool
+            Boolean specifying whether to flip vertical acceleration data before analysis (False by default). Algorithm
+            assumes that baseline vertical acceleration data is at -9.8 m/s^2 or -1g. (ie. If baseline data in vertical
+            direction is 1g, set 'flip' argument to True)
 
     '''
 
@@ -48,32 +55,40 @@ class Gaitpy():
         self.down_sample = 50
 
     def extract_features(self, subject_height, subject_height_units='centimeters', sensor_height_ratio=0.53, result_file=None, classified_gait=None, ic_prom=5, fc_prom=10):
-        ''' Continuous wavelet transform based method of gait feature detection optimization methods
+        ''' Inverted pendulum and continuous wavelet based method of gait feature detection
 
-        subject_height: int or float
-            Height of the subject. Accepts centimeters by default.
+        Parameters:
+            subject_height: int or float
+                Height of the subject. Accepts centimeters by default.
 
-        subject_height_units: str
[…]
