### 2019-10-28 matt002/GaitPy#9 @matt002
Yes that is good advice, I am working to put GaitPy on Conda-forge (see [pull request](https://github.com/conda-forge/staged-recipes/pull/9979)). In the meantime I wrote in the installation instructions that it has been tested using Py3.6 on Mac OSX (has not actually been tested on Linux or Windows). Please let me know if you are running into installation issues.
