### 2019-10-29 c456aad
Update setup.py
fix numpy bug
