### 2019-10-30 d7a4459
update requirements
update scikit-learn
