+++ requirements.txt
@@ -2,7 +2,7 @@ pandas==0.20.3
 scipy==1.2.0
 numpy==1.13.3
 PyWavelets==0.5.2
-scikit-learn==0.18.2
+scikit-learn==0.21.3
 statsmodels==0.8.0
 bokeh==0.12.10
 dill==0.2.7.1
