### 2019-10-30 matt002/GaitPy#9 @matt002
I made some updates to the setup.py and requirements.txt in order to be compatible with conda-forge. Gaitpy v1.5.5 is now passing all windows/mac/linux tests on conda-forge and I am waiting for them to merge the pull request. I have pushed these updates to Pypi as well.
