### 2019-10-30 matt002/GaitPy#9 @abhishektha
@matt002 I'm able to install Gaitpy on Windows through pip now! Thanks for getting that fixed!
