### 2019-11-01 matt002/GaitPy#9 @finsberg
It does not work on Mac, I am using python 3.7.3. Here is the output:

```
$ pip install gaitpy
Collecting gaitpy
  Downloading https://files.pythonhosted.org/packages/b9/37/b0f0e860ceeac9a7ec6c3231d74022c7e50eb3fef44f21e9a6a700ef307a/gaitpy-1.5.5-py3-none-any.whl (2.3MB)
     |████████████████████████████████| 2.4MB 2.5MB/s
Collecting statsmodels==0.8.0
  Using cached https://files.pythonhosted.org/packages/72/16/d7e7a70fc8ca3cc0d783a66e902a7adf80a810695c357cd48bb22c82451a/statsmodels-0.8.0.tar.gz
     command: /Users/henriknf/test/venv/bin/python -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/47/_3mbvrgn5gx2sqyjvqyyfbm40000gn/T/pip-install-8o5ye01z/statsmodels/setup.py'"'"'; __file__='"'"'/private/var/folders/47/_3mbvrgn5gx2sqyjvqyyfbm40000gn/T/pip-install-8o5ye01z/statsmodels/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/47/_3mbvrgn5gx2sqyjvqyyfbm40000gn/T/pip-install-8o5ye01z/statsmodels/pip-egg-info
         cwd: /private/var/folders/47/_3mbvrgn5gx2sqyjvqyyfbm40000gn/T/pip-install-8o5ye01z/statsmodels/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/47/_3mbvrgn5gx2sqyjvqyyfbm40000gn/T/pip-install-8o5ye01z/statsmodels/setup.py", line 335, in <module>
        from numpy.distutils.misc_util import get_info
    ModuleNotFoundError: No module named 'numpy'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```
