### 2019-11-01 matt002/GaitPy#9 @finsberg
And if I first install NumPy then the full traceback can be found in [this file](https://github.com/matt002/GaitPy/files/3798055/output.log)
