### 2019-11-01 #1778 @finsberg
@arfon I am still not able to install this software using pip (see https://github.com/matt002/GaitPy/issues/9). When this issue is fixed I will recommend acceptance.
