### 2019-11-01 matt002/GaitPy#9 @matt002
I am finding this error as well using Py3.7. I believe it is due to pandas, numpy, and statsmodels pinned dependencies not cooperating with Py3.7. However, unpinning these dependencies results in unit testing errors. Do you have installation troubles using Py3.6 on MacOS? I have specified in the documentation that this package is compatible with Py3.6 until I work through these issues.
