+++ docs/html/_sources/index.rst.txt
@@ -18,7 +18,7 @@ The source code is available on Github: `github.com/matt002/GaitPy <https://gith
 
 Installation
 ------------
-GaitPy has been tested using python v3.6 on Mac OSX.
+GaitPy is compatible with python v3.6 on MacOSX, Windows, and Linux.
 
 Installation via pip:
 
