### 2019-11-01 matt002/GaitPy#9 @finsberg
OK. Python3.6 works on MacOS.
