### 2019-11-01 #1778 @finsberg
@arfon I am done with my review, and recommend GaitPy for publication in JOSS.
