### 2019-11-01 #1778 @arfon
> @arfon I am done with my review, and can recommend GaitPy for publication in JOSS.

👍 many thanks @finsberg!
