### 2019-11-03 #1778 @abhishektha
@arfon I am done with my review as well and can recommend GaitPy for publication. Its a very cool package, I was able to run some of my own IMU data through it and it did a great job predicting initial and final contact times.

I did notice that one of the references was missing its DOI in the paper, I put a pull request in the repository to hopefully fit it: https://github.com/matt002/GaitPy/pull/18
