### 2019-11-03 #1778 @arfon
@mdc007 - I've also fixed a DOI in your paper here: https://github.com/matt002/GaitPy/pull/19
