### 2019-11-03 #1778 @matt002
@arfon @finsberg @abhishektha I have incorporated the above mentioned updates. Thank you all very much for your review and edits. The paper and package are much improved following the review process.
