### 2019-11-04 #1778 @matt002
@arfon I made a new release (1.6.0) and archived the software in Zenodo with DOI: 10.5281/zenodo.3528115.
